requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
Pillow==6.2.1
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
